Convert 'protocol' mod to 'net'
